vc_investment_radar/requirements.txt
beautifulsoup4==4.12.3
feedparser==6.0.11
html5lib==1.1
requests==2.32.3
PyYAML==6.0.2
